reader: reduce minimum memory requirement
Previously we wanted each thread write cache slot to be at least
processors + 1 in size to avoid a race dead-lock.   This has now
been solved with the introduction of the combined queue_cache
implementation.

However, it still makes sense to have this check, because we
potentially need each slot to be at least processors in size to
fully utilise each processor (e.g. if we're compressing a large
multi-megabyte file, we need at least 10 buffers to keep all the
deflate threads compressing that file on a 10 core/thread CPU).

So keep the check, but reduce the minimum to processors.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>