mksquashfs: write file data in separate scan before metadata creation
Separate the writing of file data and the creation of inodes into two
separate scans.  Dir_scan7() now writes file data only, and dir_scan8()
creates the file and directory inodes.

This is being done to add some parallelisation improvements to file
data writing, which cannot be done if (as at present) writing (queueing
to disk) is immediately followed by inode creation, and this has to be
done before moving on to the next file.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>