Makefile: build all compression algorithms by default
Traditionally Squashfs used GZIP, but, it has had support for
different compression algorithms since 2009, and LZO/XZ was added
in 2010, and LZ4 in 2013.  The last compression algorithm added (ZSTD)
was added in 2017.

Distribution releases (especially long term support releases) have
a long lifetime, and it be can a long time before it can be assumed
all distributions will support a "new" compression algorithm, and
thus it is difficult to know when to change something from optional to
always built by default.

But it has been a long time now.  So it now makes sense to build all the
compression algorithms by default.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>