mksquashfs: don't truncate output file when duplicate or read error occurs
Previously in the following two instances

1. A file was found to be a duplicate, and data blocks had to be
   spilled to disk during the duplicate check because the write
   cache was too small to hold the entire file.

2. A file has been found to either have I/O errors or changed size
   between stating and reading.

Mksquashfs would need to reset the current write position within the
output file to before that happened, and it truncated the output file
to that position.

But Mksquashfs doesn't need to truncate the file every time that happens.
Instead it only needs to truncate the file once the entire filesystem
has been written to ensure that there are no previous blocks beyond
the end of the filesystem.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>