virt_disk_pos.h: add concept of a virtual write position
Multiblock file writing in Mksquashfs uses the REAL output position to
send the blocks to the writer thread, and to hash the blocks in the
write cache.

Add the concept of a virtual write position and move multiblock writing
over to using that.  This will eventually mean Mksquashfs doesn't need
to know where the blocks will be written to.

Currently the virtual position tracks the real position.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>