mksquashfs: remove fragment flush stall
When a multiblock file is to be written to the output filesystem,
Mksquashfs had to first wait for all in-flight fragment blocks to be
compressed and written to disk.  This is because to make the output
filesystem reproducible all data should be written in a strict sequence,
and at the point of writing a multiblock file, the location of the
blocks in the output filesystem is unknown until any preceeding fragment
blocks has been compressed.

This commmit removes the fragment flush stall, by using virtual disk
locations at the point of writing the multiblock files, with an
"orderer" thread allocating physical locations once any preceeding
fragment blocks have been compressed.  In other words the job of
allocating physical locations is handed off to a thread to be done once
they are known, rather than stalling the main thread waiting for that
information to become available.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>