mksquashfs: remove non reproducible code
The non reproducible code was the first solution to the fragment stall
when starting to write a multiblock file.  This code worked but made the
output non reproducible.

As you may know this caused problems and so Mksquashfs abandoned it as
the default, but it could still be used if the -non-reproducible option
was specified.

Having now fixed the fragment stall in an alternative reproducible way,
this code is now obsolete and can be removed.

The options are still retained for backwards compatibility but they are
ignored.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>