Add boilerplate to automate alloc failure checking
In Linux allocation functions like malloc() never fail with out of
memory (*).  This is because the memory is not allocated at the time of
the malloc, but, only when it accessed, on a page by page basis (4096
bytes).  This is part of Linux's default overcommit policy.

It is annoying to always check for allocation failure on something that
you know will not in practice fail, and occasionly I miss it out.  So,
this commit is the start to finally automate the boilerplate code.

Currently calloc is automated.  Later commits will add malloc, strdup
etc.

(*) unless out of address space or overcommit has been disabled by
echoing 2 to /proc/sys/vm/overcommit_memory.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>