error.h: break inability to exit program
There are a number of C files where the object file (.o) is shared
between Mksquashfs and Unsquashfs.

These files need to be careful to not depend on anything which may only
be in Mksquashfs or Unsquashfs.  One such dependency is exiting.
Unsquashfs can simply call exit(1) on error, however, Mksquashfs must
call prep_exit() before exit(1).

This has hitherto made it impossible to define a macro which exits on
error, such as MEMERROR() and use it in these files.  Which meant these
files have to pass the error back up the call chain.

This commit solves this by introducing exit_squashfs(), which is
provided by both Mksquashfs and Unsquashfs.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>