threads: change thread policy to prioritise block compressor threads
With the greater parallelism available due to parallel reader threads,
Mksquashfs can often have more compressor threads active than there are
cores (either the default of all cores or specified by -processor
option).  As this is often self-defeating, Mksquashfs has a thread
limiter, which limits the number of active threads to the available
cores.

Previously when this thread limiter was introduced, multiblock file
writes stalled waiting for all in-flight fragment blocks to be
compressed.  Due to this waiting on fragment blocks, the thread limiter
prioritised fragment compression threads, to ensure fragment blocks got
compressed as quickly as possible to minimise the stall time.

The fragment block stall has now been eliminated, which means nothing in
Mksquashfs now waits on fragment block compression (except at the end of
Mksquashfs when all in-flight blocks being compressed needs to be waited
on).

The multiblock file write waits for multiblocks to be delivered by the
block compressor threads, and as such compression of these blocks are
now higher priority than fragment block compression (which nothing now
waits on).  So change the thread limiter policy to prioritise the block
compressor threads.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>