README: fix incorrect full help text count in lines
The README incorrectly stated that the number of Mksquashfs options was 263.

This was generated by:

% mksquashfs -ho ".*" | wc
    263    1897   12976

But that did not take into account line wrapping at 80 columns
(due to the output not being to a terminal).

The correct count of the number of options in Mksquashfs is given by:

% mksquashfs -ho ".*" | grep ^- | wc
    109     965    6786

However, we're actually interested in the line count of the full
help text, which is given by:

% mksquashfs -ha | wc
    389    2716   18584

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>