print_pager: Handle Busybox and Toybox less and more
Mksquashfs passes --quit-if-one-screen to less, and --exit-on-eof to more.

Unfortunately, if less and more are provided by Busybox or Toybox
less or more will fail because they don't support these options (being
light versions for embedded systems).  This is not an ideal situation.

Fortunately, the code is already writen to handle this situation in
the guise of the function determine_pager().  This function was written
to handle Debian's /usr/bin/pager (part of their alternatives system),
and /usr/bin/pager may be less or it may be more (more or less), and
Mksquashfs has to determine this to pass the right option (as above).
But it also determines if the pager doesn't look like less or more, and
so to be on the safe side no option is passed, and the pager is used
as is.

This commit refactors the code to always run determine_pager and then
decide what to do.  In this case for Busybox and Toybox it will do the
right thing and use no option.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>