mksquashfs: add inode count to progress bar
Previously inodes (metadata) were compressed whilst files were being
compressed and written to disk.  Now inodes are compressed after files
have been processed.

This means that the progress bar will hit 100% after the files have been
processed, and it be stuck at 100% for the entire time the metadata
takes to compress.  If there is a lot of metadata, then this may make it
look as if Mksquashfs has hung.

So, to avoid this, add the inode code to the progress bar, so it still
will be less than 100% when files have been processed, and it will
continue to increment whilst the metadata is being compressed and
written to disk.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>