mksquashfs: don't use sinode_count to increase progress bar
We want to know how many inodes are to be written.  For non-appending
this is current inode_no - 1.

For appending, the calculation could be based on sinode_count (the saved
inode count).  But because this includes all inodes in the previous
filesystem  including the root inode, the calculation needs to take
that into account (we re-write the root inode which means we want to
include that in the progress bar count.  So the calculation here would
be  (inode_no - 1) - sinode_count + 1, or inode_no - sinode_count.

If the -root-becomes option is used, then the calculation changes again,
this is because the sinode_count includes the old root, but does not
include the new root, and so they cancel each other out, which means
the calculation is inode_no - sinode_count - 1.

In effect 3 slightly different calculations.

This commit introduces a new variable which is setup to take the above
into account.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>