mksquashfs: handle completely sparse files
get_marked_vpos() returns the virtual disk write position when the
write_file function was entered.  This represents the virtual start
position of any data written.

However, if no data is written then there may exist no virtual position
to real disk position mapping, which will cause the virtual -> real disk
position lookup to fail.

The rare circumstance when this happens is where the multi-block file is
entirely sparse, which means no buffers are written for that file.  Thus
there is no virtual -> real disk position mapping created for that file.
Ordinarily, a later multi-block file will create a virtual -> real disk
position mapping when file blocks are written for that file.  But if the
entirely sparse file doesn't precede a multi-block file which provides
a mapping, then the virtual -> real disk position lookup will fail.

The file in this case should not be dependent on a later file
providing a mapping.  So this commit checks to see if no mapping
has been created, and asks the orderer thread to create one.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>