read_fs: create virt->disk mapping when appending
When reading and adding files already in the filesystem (for duplicate
checking), we need to add a virt->disk mapping for the disk blocks. This
mapping is one to one, where the virt value is the same as the physical
value.

The virtual and physical disk positions are set to the end of the
filesystem being appended to after the existing filesystem has been
read, and hence the above generates values which are unique and will not
clash with any later virtual values.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>