mksquashfs: extend -pd option to work for root (/)
When no sources are specified on the command line (and no -cpiostyle or
-tar), the filesystem is created entirely from pseudo definitions,
including the root directory (/).

The -pd option defines a default directory to be used when a directory
in the path of a pseudo definition is missing.  This should work for the
root directory, but it didn't, because the root directory is specially
handled.

So extend the code to allow the -pd definition to be used for the root
directory.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>