Move Hard-link printing from metadata creation time
Regular files and their hard-links were handled in the tree walk that
output them and the metadata (inodes, directories etc.).

Writing of regular file data was moved to a separate phase (dir_scan7),
but, printing of hard-links to regular files remained in the metadata
creation tree walk.

Move it to the regular file data writing phase.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>