progressbar: use 95-100% for compressing metadata
Previously, metadata was compressed while compressing files.  This meant
the time compressing metadata was included in the overall progressbar
progress from 0 - 100%.

Then metadata was moved to be done after files were compressed.  This
meant the progressbar would be at 100% for the entire time the metadata
was being compressed.  This could make someone think Mksquashfs had
hung.

I fixed this by adding the inode count to the progressbar.  But, this
typically doesn't produce a linear progressbar.

If we have 1024 files, with one block per file, this makes 1024 blocks.
Also if we have 1024 files, this makes 1024 inodes.

So the progressbar size is 2048, with 50% blocks, and 50% inodes.
But the blocks are much bigger and slower to compress than an inode.
So, in practice it means Mksquash spends 98% of the time with the
progressbar at 50% or less, and then jumps to 100%.  So the percentage
has been completely inaccurate.

But if we showed inodes at the true proportion of size/time to compress
in relation to the data, then it would less than 1%, or too small to
show up on the progressbar, and it still will be stuck at 100%.

The ideal situation would be to have a second progress bar solely for
metadata.  But that could very easily confuse any scripts out there
that expect a single progressbar incrementing from 0 - 100%.

So this commit compromises by making the metadata occupy the 95%-100%
portion of the progressbar.  This is no so great that it badly distorts
the lineararity of the progressbar, but it does exaggerate the metadata
portion so that is visible on the progressbar, and it will make
progresss rather than being stuck at 100%.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>