progressbar: switch from float to double
When a progressbar has a large amount of blocks, and the increment (inc)
is a small fraction, float has insufficient resolution.  So switch to
using double.

Also due to resolution, the incremented estimated_uncompressed count may
be one less than estimated_uncompressed at the end of Mksquashfs,
meaning the progressbar gets to 99% but not 100%.

So add a new progressbar_finish() routine which displays the progressbar
at 100%.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>