Unsquashfs: Fix one-off allocation bug in pseudo file non-printable xattr output
When generating a pseudo file of the input Squashfs filesystem, if the
xattr value had non printable values, the code failed to allocate space
for the zero terminator.

The bug manifests if the non-printable value is the last value to be
printed.

Found in final testing of the release by Valgrind memcheck.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>