tar: Fix an out of bounds memory access in pax_header() parser function
This commit fixes issue https://github.com/plougher/squashfs-tools/issues/311

A PAX extended attribute is of the following form:

<length> <keyword>=<value>

Where <length> is the full length, including the <length> field and the
terminating newline.

The <length> count should not extend beyond the end of the buffer, where
<length> is from the current parse position in the buffer.

The code has a check for <length> greater than data size (length >
st_size), which should serve as both a sanity check, and an out of
bounds check, but, the code doesn't decrease st_size to reflect the
amount of bytes left to parse.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>