mksquashfs: compute latest inode time in dir_scan6()
Previously to implement -mkfs-time inode, the latest inode time
was computed in create_inode().

This has the drawback that the latest inode time is known *after*
dir_scan8() and create_inode() has been called for all inodes.   This
is OK for computing the mkfs-time, but, not if you want to know the
latest inode time *before* dir_scan8() and create_inode() being called.

So move the computation to dir_scan6().

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>