mksquashfs: check for timestamps before epoch of 1970-01-01
Surprisingly Mksquashfs didn't have a check for files with timestamps
before the epoch of 1970-01-01, which cannot be stored in Squashfs.

But, I've never hit this issue, and I doubt m(any) users of Squashfs
have either.  But for correctness put a check in, and round any such
timestamps to 1970-01-01, with a warning.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>