mksquashfs: Don't let fabricated inodes set latest inode time
In certain cases Mksquashfs needs to fabricate a root inode (*),
and when it does it sets the timestamp to the current time.

Unfortunately, in those cases, that timestamp will likely be
the latest inode timestamp, and it will thus be choosen as the
latest inode with the -mkfs-time inode and -root-time inode options.

So avoid having these fabricated inodes setting the latest
inode time.

(*) tar files also need to fabricate leading directories if the
tar file doesn't define them.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>