mksquashfs: Use latest inode time with pseudo definitions
If a pseudo definition (standard non timestamp version) is used, then
the timestamp used for the definition will be the current time.

This fabricated time, in a similar fashion to the already fixed root
inode time (and tar default directory time), will tend to be the latest
inode timestamp, and thus will be timestamp chosen for the -mkfs-time
inode and -root-time inode options.  Which is obviously not the intended
behaviour of these options, and it will mean the filesystem will be
non-reproducible.

So in a similar fashion to the previous commit, if -mkfs-time inode or
-root-time inode is in effect, pseudo definitions don't affect the
latest inode time value, and they use the latest inode time value
as their timestamp.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>