tar: optimise read_binary() overflow check
If the size of the binary number is 7 bytes, then by
definition it cannot overflow a long long, and so
remove the unnecessary overflow check in read_binary().

Open code the 11 byte binary number read, to be valid the
first 3 bytes must be zero, and the remaining 8 bytes are
valid if they're a positive number.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>