tar: handle negative mtime header value
If a file has a negative mtime (before epoch of 1970-01-01) and GNU tar
is outputting in either "gnu" or "oldgnu" extended tar format, it will
encode the value as a binary negative number in the mtime header field.

Extended the parser to handle this negative number.

The other part of this commit is the Squashfs filesystem cannot encode
timestamps before the epoch of 1970-01-01, and so round any earlier
timestamps to 1970-01-01 with a warning.

A later commit with deal with the pax tar format.

Fixes issue https://github.com/plougher/squashfs-tools/issues/315

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>