tar: simplify read_pax_number() error checking
In all but two cases the bytes value returned by read_pax_number() is
only used to check that the character at the end of the pax number is
the expected value.

Simplify this use-case by making read_pax_number() check that the
character at the end of the pax number is the expected value.

The remaining two cases use the original read_pax_number() now renamed
to read_pax_number_ext().

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>