tar: handle negative PAX mtime value
If a file has a negative mtime (before epoch of 1970-01-01) and the tar
file format is PAX, it will encode the mtime as a PAX variable with a
negative ascii value (e.g. -1234)

Extend the parser to handle this negative number.

Additionally, the PAX value may or may not be fractional, so extend the
error checking to allow for either '.' or '\0' after the PAX value.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>