print_pager: allow quoted strings and string concatenation in PAGER
The previous commit allowed the PAGER variable to have spaces
and arguments in it.  This commit extends the parser to allow quoted
strings and to concatenate strings.

This is required if any argument in PAGER contains spaces, which means
to be passed correctly to the pager they need to be quoted, using
either double or single quotes.

For example if you wanted 'less' to copy the input to a filename with
spaces in it (e.g. "file name"), you could set PAGER to

% export PAGER='less "--log-file=file name"'

Without support for quoted strings, 'less' will fail with

-----
"--log-file=file: No such file or directory
name": No such file or directory
Pager failed to run or failed with an error status
-----

String concatenation is also required because often people will quote
the filename and not the argument itself, e.g.

% export PAGER='less --log-file="file name"'

here "--log-file=" has to be concatenated with "file name" to produce
"--log-file=file name".

Obviously, this allows multiple strings to be specified, e.g.

% export PAGER='less "--log-file"="file name"'

such a situation is unlikely to be deliberate, but, can be produced by
the expansion of variables with embedded quotes, e.g.

$ DIR='"/tmp"'
$ FILE='"file name"'
$ export PAGER="less --log-file"=$DIR/$FILE
$ echo $PAGER
less --log-file="/tmp"/"file name"

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>