print_pager: update error messages regarding shell special chars in PAGER
PAGER cannot have shell special characters as they will not be treated
specially and instead they will be passed to the pager command which
will most likely fail in an unexpected or obscure way.

As a courtesy the parser checked for these, and told the user they were
disallowed.

Update the error messages to be more explicit about what the parser
is complaining about, and update the advice in that if you really
want to pass these characters to the pager, you can now do that by
quoting or backslashing the characters.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>