mksquashfs: ignore -no-pager in main command line parser loop
The obvious use of -no-pager is to disable the pager when running one of
the help options (e.g. mksquashfs -no-pager -help-all).  This will make
Mksquashfs exit before running the main command line parser loop.

However, if -no-pager is entered on it's own (with source/destination
e.g. mksquashfs a b -no-pager), the main loop will not recognise
-no-pager as a valid option.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>