mksquashfs: fix -offset regression
Issue https://github.com/plougher/squashfs-tools/issues/319 reports
that -offset (-o) is broken in Mksquashfs 4.7.

This is caused because version 4.7 always explictly truncates the output
file to the filesystem length at the end of writing the filesystem.
This is remove any extraneous data spilled to the output filesystem
in duplicate checking (if a file cannot be entirely held in memory,
parts of it have to be spilled to disk, which then become extraneous
if the file is found to be a duplicate).

Unfortunately the file is truncted to the filesystem size without the filesystem
offset.

This commit fixes https://github.com/plougher/squashfs-tools/issues/319

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>