sqfstar: fix -offset regression
-offset (-o) is broken in Sqfstar 4.7.

This is caused because version 4.7 always explictly truncates the output
file to the filesystem length at the end of writing the filesystem.
This is remove any extraneous data spilled to the output filesystem
in duplicate checking (if a file cannot be entirely held in memory,
parts of it have to be spilled to disk, which then become extraneous
if the file is found to be a duplicate).

Unfortunately the file is truncted to the filesystem size without the filesystem
offset.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>