mksquashfs: don't memmove() in get_inode() for every 8K block
Compressing a large sparse file (256 Gbytes) using a 4K block size, it
was discovered Mksquashfs was hanging in get_inode().  This is because
the routine unnecessarily memmoves() the cached inode data for every 8K
compressed, and given that the cached data starts out at 256 Mbytes,
that is a lot of memmoving!

Also use memcpy() rather than memmove() on the remaining sub-8K data,
as the source and destination doesn't overlap.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>