mksquashfs: don't wait for all data blocks to go to orderer thread
Before the file inodes are created, the on disk location of the data
blocks need to be known.

Previously, Mksquashfs stalled waiting for all blocks to be processed
by the orderer thread before creating the file inodes.

This, however, creates the situation that metadata compression is
unnecessarily stalled waiting on fragment blocks and data blocks that
don't need to be waited on (i.e we don't need to wait for last data
block to go to disk before compressing the inode for the first data
block).

This commit removes the stall, and instead when the disk block location
is needed in the file inode creation, it is waited on then if it is not
already available.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>