actions: add align() action to set file alignment
This commit adds the action align(<alignment>) which will align any file
which matches the test operator(s) to <alignment>.  Alignment should be
a pure power of 2 and 64 Megabytes or less.

Any file which has an alignment applied will be separately compressed,
and *not* packed into a fragment block.  This may reduce compression.

Note due to Squashfs filesystem layout restrictions, the alignment is
applied to the first block in a file only.  Therefore if you want
all data blocks to be aligned you should use a block size large enough
to accommodate the largest file you want aligned.

Some Actions examples for those unfamiliar with the Actions
functionality.

% mksquashfs dir dir.sqfs -action "align(4096)@true"

This will align all files to a 4096 byte boundary

% mksquashfs dir dir.sqfs -action "align(4K)@filesize(>=32K)"

This will align all files which are 32K or larger to a 4K boundary.

% mksquashfs dir dir.sqfs -action "align(4K)@perm(/a=x) && user(root)"

Align files which are executable and owned by root, to a 4K boundary.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>