mksquashfs: allow streaming where filesystem is output to stdout
I have often wanted the ability to stream the output of Mksquashfs,
where the filesystem is output to stdout, rather than a file.  This
obviously allows the output to be streamed over ssh, and is useful
when the system running Mksquashfs doesn't have sufficient disk
space to hold the output filesystem.

Other people obviously also would like this facility, and it was
raised as an issue back in 2020

https://github.com/plougher/squashfs-tools/issues/88

As I mentioned in that issue, Mksquashfs wants the output file
to be seekable (which stdout isn't) because it has to write the
Squashfs super block to the start of the file when it finishes.

It also needs the output filesystem to be seekable to do duplicate
checking.

But, if you can live without duplicate checking, it is possible to
output the filesystem to stdout, by writing the super block to the
end of the file.

The resultant file won't be a proper Squashfs filesystem because the
super block is at the end of the file.  But, it can be fixed up
afterwards with a Mksquashfs option which simply writes the super
block to the start of the file.  In addition Unsquashfs can be
taught to look for the super block at the end of the file (for a
streamed file) and so it will happily work on an unfixed up streamed
file.

This commit adds the -stream option which will output the filesystem
to stdout, and put the super block at the end of the file.  For
obvious reasons duplicate checking is disabled, and anything that
is normally output to stdout, like the progress bar and the summary
is disabled too.

To work with the existing Mksquashfs command line parser, the output
file has to be given, but it will be ignored, e.g.

% mksquashfs dir - -stream

will output the filesystem to stdout.

For those technically minded, the Squashfs super block at the start
of the file is written zeroed out, with the special Squashfs magic
HSQS.  This is used by Unsquashfs to look for the super block at the
end of the file.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>