mksquashfs: add -fix option
This commit adds the -fix option, which can be used to fix up a Squashfs
filesystem streamed to stdout by Mksquashfs.

The -fix option basically checks the filesystem is a "streamed Squashfs"
filesystem, and then writes the super block at the end of file to the
start of the file, and then truncates the file to remove the super
block at the end of the file.

Usage is:

% mksquashfs -fix file

This commit is one of three which fixes issue
https://github.com/plougher/squashfs-tools/issues/88

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>