pseudo: fix hang in parallel pseudo read files
I have spent the last couple of days tracking down and fixing a
difficult bug when pseudo read files (reading from a command) are
run in parallel on multiple reader threads.

When reading output from a command (zstd -dc), occasionally a pseudo
read file will hang.  This is because the parent reader thead has hung
in read(), even though the child process has died, and thus the pipe
should be closed and the read() should have returned EOF.

In more detail, the standard way to execute a command in a sub-process
is to:

1. Create a pipe
2. Call fork()
3. The child process will:
	1. redirect STDOUT to the write end of the pipe (pipefd[1]),
	2. exec the command.
4. The parent process will:
	1. close the write end of the pipe (pipefd[1]),
	2. repeatedly call read() on the read end of the pipe, until
	   it returns EOF.
	3. call waitpid() to get the exit status.

Once the child process dies and closes the pipe, the read() (4.2
above) should finish, but it is hanging.

Seaching the collected wisdom of the internet, a common notion is to
blame this on the fact that the child process will be a zombie (defunct
progress), because it has not been waited on because the parent process
is stuck in the read().  In other words the read() won't exit until
waitpid() is called, and waitpid() won't be called.

See

https://stackoverflow.com/questions/53556070/read-hangs-on-zombie-process
https://stackoverflow.com/questions/53562111/how-to-avoid-read-from-hanging-in-the-following-situation

The "solution" given on the internet is to do the waitpid()
asynchronously on a signal (SIGCHLD), see

https://stackoverflow.com/questions/40601337/what-is-the-use-of-ignoring-sigchld-signal-with-sigaction2

But, there are (in my mind) two rather obvious issues with the above
stated causes, and a demonstratively obvious issue with the "solution":

	1. A zombie progress is dead, defunct.  Any pipe closing will
	   have already happened.  Pipe closing doesn't happen once the
	   process has been reaped.

	2. If you had to call waitpid() before read() exits, then it
	   will never work ... not just occasionally not work.

	3. The solution doesn't work.  In my case the read() still
	   hangs but the child process has died, and it now doesn't
	   exist as a zombie or in any other state.  Ergo it shows
	   zombie processes are not the issue.

The above is a classic case of confusing a symptom (zombie process)
with the cause, and then fixing the symptom.  In all likelihood for
many people moving to doing waitpid() asynchronously changes the
scheduling so that the proplem goes away, but the fundamental cause
doesn't.

So what is actually going on here?  We go back to the basic concept
which is when the child dies, the write end of the pipe closes, and
the read() exits with EOF.  So if that hasn't happened, and we *know*
the child process has died and closed the pipe, then the cause must be
*other* processes have the write end of the pipe open.

You should then observe that threads share the same file descriptor
table, and so a pipe opened in one thread will show up in another
thread.

So if multiple threads call pipe() and fork() simultaneously, a child
process will not only inherit its own pipe, but, it will also inherit
other processes' pipes, and this is before the parent processes have
closed the write end of those pipes.  In otherwords, you end up with
multiple processes with the write end of the pipe open.

Inspecting /proc/xxxxx/fd on the child processes indeed show they
have inherited multiple pipes with the write end open.

But having multiple processes with the write end of the pipe open isn't
actually enough to cause read() to hang.  The processes have to also be
inter-dependant in some way due to that shared pipe that causes them to
mutually deadlock each other.  That is if all the processes are
independant, they will all eventually die closing their write end of the
inadvertantly inherited pipe, and so the read() will exit albeit later
than planned.

So what is making these processes which have the inherited pipe with the
write end open to mutually deadlock?  It is because their output (via
STDOUT and the pipe) to their reader thread is larger than the reader
thread buffer space available, and so they go to sleep waiting for a
free buffer in the reader thread.

The following example should make it clearer:

1. Child process A gets created to execute "cat file A", and it
   inherits the open write end of pipe A.

2. Simultaneously, child process B gets created to execute "cat file B",
   and it inherits open write ends of *both* pipe A and pipe B.

3. Child A exits closing its write side of pipe A.

4. Child B goes to sleep outputting to reader thread B (via STDOUT
   and the pipe) because reader thread B is out of buffer space.

Now the read() on Child A won't exit because Child B has the write
end of pipe A open.  Child B won't exit closing the write side of
pipe A because reader thread B can only get free buffers when the
occupied buffers are written to disk, and that will only be when
Mksquashfs starts writing file B, which is once file A has finished
being read, and the read() for file A is stuck because Child B has the
write end of the pipe open.  Deadlock.

The solution is to prevent child processes from inadvertantly
inheriting the open write end of pipes which don't belong to them.
This can be done by making the "pipe(), fork() and close(pipefd[1])"
sequence atomic, and so only the child for which it was intended gets
the write end of the pipe open.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>