mksquashfs: add '-min-sparse-copy <size>' option
This allows instead of copying full contents into an image to put a sparse file of the same size for all files which are atleast of size <size>. E.g. when using '-min-sparse-copy 1' the image contains only files with '0' bytes as contents and their original file size. For 'min-sparse-copy 1000' all files with atleast 1000 bytes contain only '0' as contents.