mksquashfs: fix regression introduced by -stream option
When using -stream the super-block is written *after* the end of the
fileystem, and after any padding bytes.  This is to make fixing the
filesystem some trivial write() to block 0 and trucate() calls.

However it introduced the side-effect of leaving the write position
(get_dpos_and_inc()) at the end of the padding bytes (rather than
previously before), which means the filesystem size in the stats summary
subtly changed because it used get_dpos() to get the size of the
filesystem which is now wrong (the size shoud not include the padding
bytes).

Fix this by using the bytes_used field in the super-block which is
updated before the padding bytes are added.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>