sqfstar: add -numeric-owner option
Add -numeric-owner option which tells Sqfstar to use the numeric UID and
GID in the tarfile, rather than any user name or group name.

This is useful where the user name or group name maps to a different UID
or GID on the system where Sqfstar is being run.

This option is based on the GNU tar --numeric-owner option.

Fixes https://github.com/plougher/squashfs-tools/issues/285

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>