mksquashfs: add -numeric-owner option
Add -numeric-owner option which tells Mksquashfs to use the numeric UID
and GID in the tarfile, rather than any user name or group name.

This is useful where the user name or group name maps to a different UID
or GID on the system where Mksquashfs is being run.

Fixes https://github.com/plougher/squashfs-tools/issues/285

This option is based on the GNU tar --numeric-owner option.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>